clase Admin

+++ app/src/main/java/com/lscampov/firebase/Member.java
@@ -2,9 +2,9 @@ package com.lscampov.firebase;
 
 public class Member {
     private String name;
-    private Integer age;
-    private Long phone;
-    private Float ht;
+    private double age;
+    private double phone;
+    private double ht;
 
     public Member() {
 
@@ -18,27 +18,27 @@ public class Member {
         this.name = name;
     }
 
-    public Integer getAge() {
+    public double getAge() {
         return age;
     }
 
-    public void setAge(Integer age) {
+    public void setAge(double age) {
         this.age = age;
     }
 
-    public Long getPhone() {
+    public double getPhone() {
         return phone;
     }
 
-    public void setPhone(Long phone) {
+    public void setPhone(double phone) {
         this.phone = phone;
     }
 
-    public Float getHt() {
+    public double getHt() {
         return ht;
     }
 
-    public void setHt(Float ht) {
+    public void setHt(double ht) {
         this.ht = ht;
     }
 }
